+++ src/helpers/functions/Game.js
@@ -1,5 +1,5 @@
 const { gameStatuses, gameDefaultOptions } = require("../constants/Game");
 
-exports.getGameStatuses = JSON.parse(JSON.stringify(gameStatuses));
+exports.getGameStatuses = () => JSON.parse(JSON.stringify(gameStatuses));
 
-exports.getGameDefaultOptions = JSON.parse(JSON.stringify(gameDefaultOptions));
+exports.getGameDefaultOptions = () => JSON.parse(JSON.stringify(gameDefaultOptions));